requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
googlemaps==4.10.0
python-dotenv==1.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
